Add optional data when issuing an Asset
This commit adds a UTF-8 data field attachable when calling issueAsset API.
The data is retrievable in all APIs which return the full asset object response.
Additionally, this adds the ability to get the asset.data when calling
getAllOpenAskOrders with includeAssetData = true.

This commit does not expose these changes to the UI.